requirements.txt
appdirs==1.4.4
astroid==2.1.0
distlib==0.3.1
filelock==3.0.12
importlib-metadata==2.0.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==8.1.0
pylint==2.2.2
rope==0.18.0
Shapely==1.7.1
six==1.12.0
virtualenv==20.1.0
wrapt==1.11.1
zipp==3.4.0
